% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedition_tools.R
\name{backup_data}
\alias{backup_data}
\title{Backup Data}
\usage{
backup_data(list = NULL, file_path = NULL)
}
\arguments{
\item{list}{The list returned from \code{\link[=write_data]{write_data()}}}

\item{file_path}{The file path to the backup folder}
}
\value{
backup The status of the backup.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function backs up the data that has just been written from \code{\link[=write_data]{write_data()}}
}
