% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_functions.R
\name{body_builder}
\alias{body_builder}
\title{Build the body of an HTML email}
\usage{
body_builder(title = NULL, sections = NULL)
}
\arguments{
\item{title}{metadata title of the email}

\item{sections}{Pass the sections of the email into a vector. If you include 5 elements, then 5 sections will be created with a line between each. If you'd like HTML
formatting in the sections, i.e. bolding text, the character text needs to be HTML formatted}
}
\description{
It's common in the Crosbie lab to need to send an email as a part of an automated pipeline. This function reduces the time to create and HTML body.
}
\details{
The builder automatically adds the formatting HTML like <tr>, etc. Use the sections param to create the sections in the email. This assumes everything is left aligned.
For more complicated bodies, you'll need to build it yourself or update this function
}
