% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedition_tools.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(list = NULL)
}
\arguments{
\item{list}{a list that must contain the following
\itemize{
\item \code{df}: The data in the associated REDCap project
\item \code{dd}: A dataframe with the project's data dictionary
\item \code{events}: A dataframe with the project's event mapping
\item \code{info}: A dataframe with the project's info API call
}}
}
\value{
A list of dataframes
\itemize{
\item \code{df}: The data in the associated REDCap project
\item \code{dd}: A dataframe with the project's data dictionary
\item \code{events}: A dataframe with the project's event instrument mapping
\item \code{info}: A dataframe with the information about the REDCap project
}

@seealso \code{\link[=clean_checkboxes]{clean_checkboxes()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function "cleans" REDCap data from the \code{\link[=get_data]{get_data()}} function. To clean the data, the function uses the \code{\link[=clean_checkboxes]{clean_checkboxes()}} function and the project's data dictionary to expand the checkboxes
and the form statuses to change 0's to NA when an instrument is empty.
}
