% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tools.R
\name{clean_export}
\alias{clean_export}
\title{Clean Export}
\usage{
clean_export(df = NULL)
}
\arguments{
\item{df}{A dataframe to be cleaned}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Data exported from Expedition have standardized SQL notations. Data from REDCap also have standaradized structures.
This function cleans for comparison with other data.
}
