% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_dictionary_functions.R
\name{clean_redcap_dd_names}
\alias{clean_redcap_dd_names}
\title{Clean REDCap DD Column Names}
\usage{
clean_redcap_dd_names(dd = NULL)
}
\arguments{
\item{dd}{a REDCap data dictionary}
}
\value{
REDCap data dictionary with corrected names
}
\description{
REDCap data dictionaries exported via the API return different column names compared to
data dictionaries manually exported from a file. This function checks which source was used to export a
data dictionary and harmonizes the column names.
}
