% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_dictionary_functions.R
\name{clean_redcap_statuses}
\alias{clean_redcap_statuses}
\title{Clean REDCap Status Questions}
\usage{
clean_redcap_statuses(df = NULL, dd = NULL)
}
\arguments{
\item{df}{The redcap export to clean}

\item{dd}{The data dictionary that corresponds to the redcap export to clean}
}
\value{
A redcap export with cleaned status questions

Sometimes questions have been moved to inactive instruments
When this occurs, this method for cleaning checkboxes doesn't work, so I make sure no 1's are overwritted.
}
\description{
When data are exported from REDCap via the API, the status of each form is also exported. These status questions are always formatted as [form_name]_complete.
These fiels default to 0 in the export even if the form hasn't been opened yet. If a form hasn't been opened, we want these questions to be NA.
This function addresses this issue
}
