% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tools.R
\name{compare_df}
\alias{compare_df}
\title{Compare the the differences between two dataframes}
\usage{
compare_df(input_df = NULL, output_df = NULL, names = c("input", "output"))
}
\arguments{
\item{input_df}{data.frame object before changes}

\item{output_df}{data.frame object after changes}

\item{names}{suffix to attach to the output data.frame. Defaults to 'input' and 'output'}
}
\value{
table with t-scores attached to raw swan values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \cr
Commonly in the Crosbie lab, we need to be visually inspecting the changes code makes to a dataframe. This function takes an input and an output dataframe and returns the differences between the two
}
