% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{compare_dfs}
\alias{compare_dfs}
\alias{compare_df}
\title{Compare the the differences between two dataframes}
\usage{
compare_df(input_df = NULL, output_df = NULL, names = c("input", "output"))
}
\arguments{
\item{input_df}{data.frame object before changes}

\item{output_df}{data.frame object after changes}

\item{names}{suffix to attach to the output data.frame. Defaults to 'input' and 'output'}
}
\value{
table with t-scores attached to raw swan values
}
\description{
Commonly in the Crosbie lab, we need to be visually inspecting the changes code makes to a dataframe. This function takes an input and an output dataframe and returns the differences between the two
}
