% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosbieLabFunctions-package.R
\docType{package}
\name{CrosbieLabFunctions-package}
\alias{CrosbieLabFunctions}
\alias{CrosbieLabFunctions-package}
\title{CrosbieLabFunctions: Commonly used functions in the Crosbie Lab to make data analysis more efficient}
\description{
A lot of work is repeatable and can be "functionized." This is a package of the functions that are commonly used in the Crosbie Lab.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jclutton/CrosbieLabFunctions}
  \item Report bugs at \url{https://github.com/jclutton/CrosbieLabFunctions/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Clutton \email{jonathan.clutton@sickkids.ca}

}
\keyword{internal}
