% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{describe_df}
\alias{describe_df}
\title{Return descriptive statistics about all variables in a data.frame}
\usage{
describe_df(df = NULL)
}
\arguments{
\item{df}{a data.frame object}
}
\value{
a data.frame with descriptive statistics
}
\description{
Return descriptive statistics about all variables in a data.frame
}
