% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscore_functions.R
\name{get_cis_tscores}
\alias{get_cis_tscores}
\title{Run analysis on Cognitive Impairment Screener (CIS) Scores}
\usage{
get_cis_tscores(df = NULL)
}
\arguments{
\item{df}{The data frame with p_respondent, age, and cis_pro}
}
\value{
Additional column of CIS tscores
}
\description{
get_cis_tscores() returns a set of t-scores for the Cognitive Impairment Screener (CIS) assessment
}
\section{Development}{

2025-08-05: Added so that could be added to S2 scripts. JC
}

