% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedition_tools.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(
  token = NULL,
  redcap_uri = NULL,
  fields = NULL,
  export_survey_fields = TRUE
)
}
\arguments{
\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{fields}{An array, where each element corresponds to a desired project
field.  Optional.}

\item{export_survey_fields}{A boolean that specifies whether to export the
survey identifier field (e.g., 'redcap_survey_identifier') or survey
timestamp fields (e.g., instrument+'_timestamp').
The timestamp outputs reflect the survey's \emph{completion} time
(according to the time and timezone of the REDCap server.)}
}
\value{
A list of dataframes
\itemize{
\item \code{df}: The data in the associated REDCap project
\item \code{dd}: A dataframe with the project's data dictionary
\item \code{events}: A dataframe with the project's event instrument mapping
\item \code{info}: A dataframe with the information about the REDCap project
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function uses a REDCap API token to return the project's data and dictionary.
This work relies on REDCapR. See \link[REDCapR]{redcap_read}
}
