% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedition_tools.R
\name{reconcile}
\alias{reconcile}
\alias{reconcile_data}
\title{Reconcile}
\usage{
reconcile_data(
  list = NULL,
  token = NULL,
  redcap_uri = NULL,
  transfer_to_braincode = TRUE
)
}
\arguments{
\item{list}{The list returned from \code{\link[=clean_data]{clean_data()}}}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{transfer_to_braincode}{T/F - True gives additional warnings}
}
\value{
A List
\itemize{
\item \code{df}: The data in the associated REDCap project
\item \code{dd}: A dataframe with the project's data dictionary
\item \code{events}: A dataframe with the project's event instrument mapping
\item \code{info}: A dataframe with the information about the REDCap project
\item \code{reconciled}: A message whether the reconciliation was successful - success or no changes
\item \code{changes}: A dataframe with the changes to be pushed
\item \code{downstream_name}: The name of the downstream project
\item \code{downstream_pid}: The project id of the downstream project
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
We now have a cleaned version of the data in \code{\link[=clean_data]{clean_data()}}. We need to reconcile for changes in the out project. This function 1) compares the metadata and 2) looks for changes in the
input to output projects using the \code{\link[=track_changes]{track_changes()}} function. If data are being overwritten, the process is stopped and a record of the differences is saved in the error_data.csv file. In order for the function to work you pass a list with
the input df, dd and events, and pass the output token and redcap_uri. The function gets the most up to date data from the output project and compares it with the input data
}
\seealso{
\code{\link[=track_changes]{track_changes()}}
}
