% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_functions.R
\name{send_email}
\alias{send_email}
\title{Send an Email}
\usage{
send_email(
  host = "smtp.sickkids.ca",
  port = 25,
  from = NULL,
  to = NULL,
  subject = NULL,
  body = NULL,
  attachments = NULL
)
}
\arguments{
\item{host}{The host of the server}

\item{port}{The port for email sending}

\item{from}{Email address of person sending email}

\item{to}{The email address or email addresses of people receiving email}

\item{subject}{Subject header}

\item{body}{The body of the email, which will be in HTML}

\item{attachments}{paths to one or multiple files to be included}
}
\value{
Success of whether email is sent
}
\description{
It's common in the Crosbie lab to need to send an email as a part of an automated pipeline. R does not provide great functionality for doing so.
This function uses the reticulate package to utilize Python and the smtplib Python package to send emails.
}
