% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tools.R
\name{uat_redcap_data}
\alias{uat_redcap_data}
\title{User Acceptance Testing for REDCap Data}
\usage{
uat_redcap_data(
  redcap_file_path = NULL,
  expedition_file_path = NULL,
  redcap_df = NULL,
  expedition_df = NULL
)
}
\arguments{
\item{redcap_file_path}{The pathway to the downloaded REDCap data, as a .csv or .xlsx file}

\item{expedition_file_path}{The pathway to the downloaded Expedition data, as a .csv file. \cr
\itemize{
\item Note: Expedition limits downloads to 100,000 rows}}

\item{redcap_df}{A dataframe with REDCap data. Use redcap_df instead of redcap_file_path if you are importing REDCap data in a different way, i.e. API.}

\item{expedition_df}{A dataframe with Expedition data. Use expedition_df instead of expedition_file_path if you are importing Expedition data in a different way, i.e. API.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
Data being uploaded into Expedition go through User Acceptance Testing. This involves ensuring that the data within Expedition are as expected.
This function is specifically designed for User Acceptance Testing \strong{REDCap} data. It allows you to compare data exported from REDCap with data exported from Expedition.\cr \cr
\emph{\strong{Instructions}}\cr
\enumerate{
\item Export raw data from REDCap
\item Export data from Expedition
\item Use this function to compare the two tables
}
}
