% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expedition_tools.R
\name{write}
\alias{write}
\alias{write_data}
\title{Write Data}
\usage{
write_data(
  list = NULL,
  token = NULL,
  redcap_uri = NULL,
  overwrite_blanks = FALSE
)
}
\arguments{
\item{list}{The list returned from \code{\link[=reconcile_data]{reconcile_data()}}}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{overwrite_blanks}{T/F - allows overwriting blanks if changing data from a value to a blank}
}
\value{
A List
\itemize{
\item \code{status}: The status of the data transfer
\item \code{datetime}: Datetime transfer occured
\item \code{project}: The name of the downstream project
\item \code{pid}: The pid of the downstream project
\item \code{server}: The redcap_uri of the downstream project
\item \code{changes}: A dataframe with the changes pushed
\item \code{df}: The data transferred
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function takes the list from reconcile data and writes data to a REDCap project
}
\seealso{
REDCapR::\code{\link[=redcap_write]{redcap_write()}}
}
