% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_dictionary_functions.R
\name{clean_checkboxes}
\alias{clean_checkboxes}
\title{Clean Checkboxes}
\usage{
clean_checkboxes(df = NULL, dd = NULL)
}
\arguments{
\item{df}{The redcap export to clean}

\item{dd}{The data dictionary that corresponds to the redcap export to clean}
}
\value{
A redcap export with cleaned checkboxes

Checkbox data always export from REDCap as either 0 or 1, regardless of if it's missing or not
A lot of the 0s should actually be NA, as the participant didn't complete the form
They are assigned a reference variable from the REDCap instrument that contained the checkbox
If the reference variable is missing, the checkbox is also set to missing
Sometimes questions have been moved to inactive instruments
When this occurs, this method for cleaning checkboxes doesn't work, so I make sure no 1's are overwritted.
Sometimes questions have been moved to inactive instruments
When this occurs, this method for cleaning checkboxes doesn't work, so I make sure no 1's are overwritted.
Sometimes questions have been moved to inactive instruments
When this occurs, this method for cleaning checkboxes doesn't work, so I make sure no 1's are overwritted.
Sometimes questions have been moved to inactive instruments
When this occurs, this method for cleaning checkboxes doesn't work, so I make sure no 1's are overwritted.
}
\description{
When data are exported from REDCap, checkboxes always export as 1 or 0 regardless of whether the question was completed.
This function corrects the check box data that should be missing by reading in the dataframe of interest and the REDCap data dictionary.
It checks for whether any questions were completed on a form. If they were, it leaves the question as is. If no questions were completed, it changes to NA.
}
\section{Important}{

This project was developed for the Braincode API Pipeline. As such we prioritized cleaning projects with longitudinal and repeating events. This function will not work properly for...
\itemize{
\item \verb{Repeating Instruments Only} - when data are not logitudinal and not repeating
\item \verb{Longitudinal with data in the repeat_instrument column } - the current iteration will not work to clean when there are repeating instruments only
This function is likely to have issues. Watch closely.
}
}

\section{Development}{

2025-09-16: Added the capacity to handle pre-expanded checkbox values using the \code{\link[REDCapR:redcap_variables]{REDCapR::redcap_variables()}} function
2025-11-24: Handled blank 'complete' forms. Not perfect. Made a work around for the expedition_tools where any form with only a complete status is removed. Might need to revisit.
}

