% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{describe_df}
\alias{describe_df}
\title{Return descriptive statistics about all variables in a data.frame}
\usage{
describe_df(df = NULL)
}
\arguments{
\item{df}{a data.frame object}
}
\value{
a data.frame with descriptive statistics
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function returns data type, count, number missing, mean, median, min, max, and standard deviation for all numberic columns in a dataframe.
It returns data type, count, and missing for any non-numeric columns in a dataframe.
}
