% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{track_changes}
\alias{track_changes}
\title{Track Changes}
\usage{
track_changes(df_in = NULL, df_out = NULL, identifier_column = NULL)
}
\arguments{
\item{df_in}{The dataframe before the changes}

\item{df_out}{The dataframe after the changes}

\item{identifier_column}{The identifier column in the data.frame}
}
\value{
A data frame with changes
}
\description{
This function can be used to see changes between two data frames. It's intended to be used to see the changes code introduces.
}
