% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{track_changes}
\alias{track_changes}
\title{Track Changes}
\usage{
track_changes(
  df_in = NULL,
  df_out = NULL,
  identifier_column = NULL,
  names = NULL
)
}
\arguments{
\item{df_in}{The dataframe before the changes}

\item{df_out}{The dataframe after the changes}

\item{identifier_column}{The identifier column in the data.frame}

\item{names}{A length of two vector for renaming the 'values' columns. Pass c('df1', 'df2') to names, where df1 = df_in and df_out = df2 to rename the output columns to match the comparison dataframes}
}
\value{
A data frame with changes
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function can be used to see changes between two data frames. It's intended to be used to see the changes code introduces.
}
\section{Development Notes}{

2025-09-12: Added naming functionality
}

